+++ Tree_Controller/Properties/Stable/ChangeLog.md
@@ -1,3 +1,6 @@
-﻿# Patch v.1.7.2
-* Fixed Color Variation options when mod could not find ColorGroupID for Seasons when the list of ColorGroupsIDs has a different order.
-* Changed default location and name of settings file. Existing settings file stays put and keeps working, but new settings file will be created at .../ModsSettings/yenyang/Tree_Controller.coc  
+﻿# Patch v.1.7.3
+* Added Min and Max Slope Limits developed by Winter.
+* Fix for not being able to save custom sets without the Advanced Forest Brush menu open.
+* Added a compatibility method for future collaboration with Find It. 
+* Fix not disposing of a persistant native list.
+* Updated Localization